+++ consensus/Cargo.toml
@@ -25,5 +25,3 @@ thiserror = "1"
 base64 = "0.13"
 time = {version = "0.3", features = ["macros"]}
 wasm-bindgen-test = "0.3"
-
-
