+++ crypto/src/hash/mod.rs
@@ -1,6 +1,7 @@
 // Copyright 2020 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod poseidon;
 pub mod prefixes;
 mod traits;
 mod types;
